numcodecs

Enum AnyArrayBase

#[non_exhaustive]
pub enum AnyArrayBase<T: AnyRawData> {
    U8(ArrayBase<T::U8, IxDyn>),
    U16(ArrayBase<T::U16, IxDyn>),
    U32(ArrayBase<T::U32, IxDyn>),
    U64(ArrayBase<T::U64, IxDyn>),
    I8(ArrayBase<T::I8, IxDyn>),
    I16(ArrayBase<T::I16, IxDyn>),
    I32(ArrayBase<T::I32, IxDyn>),
    I64(ArrayBase<T::I64, IxDyn>),
    F32(ArrayBase<T::F32, IxDyn>),
    F64(ArrayBase<T::F64, IxDyn>),
}

Numeric n-dimensional arrays with dynamic shapes.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

U8(ArrayBase<T::U8, IxDyn>)

U16(ArrayBase<T::U16, IxDyn>)

U32(ArrayBase<T::U32, IxDyn>)

U64(ArrayBase<T::U64, IxDyn>)

I8(ArrayBase<T::I8, IxDyn>)

I16(ArrayBase<T::I16, IxDyn>)

I32(ArrayBase<T::I32, IxDyn>)

I64(ArrayBase<T::I64, IxDyn>)

F32(ArrayBase<T::F32, IxDyn>)

F64(ArrayBase<T::F64, IxDyn>)

Implementations

impl<T: AnyRawData> AnyArrayBase<T>

pub fn len(&self) -> usize

Returns the total number of elements in the array.

pub fn is_empty(&self) -> bool

Returns whether the array has any elements.

pub const fn dtype(&self) -> AnyArrayDType

Returns the dtype of the array.

pub fn shape(&self) -> &[usize]

Returns the shape of the array.

pub fn strides(&self) -> &[isize]

Return the strides of the array.

pub const fn as_typed<U: ArrayDType>( &self, ) -> Option<&ArrayBase<U::RawData<T>, IxDyn>>

Returns the U-typed array in Some(_) iff the dtype of U matches the dtype of this array. Returns None otherwise.

pub fn as_typed_mut<U: ArrayDType>( &mut self, ) -> Option<&mut ArrayBase<U::RawData<T>, IxDyn>>

Returns the U-typed array in Some(_) iff the dtype of U matches the dtype of this array. Returns None otherwise.

impl<T: AnyRawData> AnyArrayBase<T>
where T::U8: Data, T::U16: Data, T::U32: Data, T::U64: Data, T::I8: Data, T::I16: Data, T::I32: Data, T::I64: Data, T::F32: Data, T::F64: Data,

pub fn view(&self) -> AnyArrayView<'_>

Returns a read-only view of the array.

pub fn cow(&self) -> AnyCowArray<'_>

Returns a copy-on-write view of the array.

pub fn into_owned(self) -> AnyArray

Turns the array into a uniquely owned array, cloning the array elements if necessary.

pub fn as_bytes(&self) -> Cow<'_, [u8]>

Returns the array’s data as a byte slice.

If the array is contiguous and in standard order, i.e. if the element order in memory corresponds to the logical order of the array’s elements, a view of the data is returned without cloning.

Otherwise, the data is cloned and put into standard order first.

impl<T: AnyRawData> AnyArrayBase<T>
where T::U8: DataMut, T::U16: DataMut, T::U32: DataMut, T::U64: DataMut, T::I8: DataMut, T::I16: DataMut, T::I32: DataMut, T::I64: DataMut, T::F32: DataMut, T::F64: DataMut,

pub fn view_mut(&mut self) -> AnyArrayViewMut<'_>

Returns a read-write view of the array.

pub fn with_bytes_mut<O>(&mut self, with: impl FnOnce(&mut [u8]) -> O) -> O

Provides access to the array’s data as a mutable byte slice.

If the array is contiguous and in standard order, i.e. if the element order in memory corresponds to the logical order of the array’s elements, a mutable view of the data is returned without cloning.

Otherwise, the data is cloned and put into standard order first, and later copied back into the array.

pub fn assign<U: AnyRawData>( &mut self, src: &AnyArrayBase<U>, ) -> Result<(), AnyArrayAssignError>
where U::U8: Data, U::U16: Data, U::U32: Data, U::U64: Data, U::I8: Data, U::I16: Data, U::I32: Data, U::I64: Data, U::F32: Data, U::F64: Data,

Perform an elementwise assignment to self from src.

Errors

Errors with

AnyArrayAssignError::DTypeMismatch if the dtypes of self and src do not match

AnyArrayAssignError::ShapeMismatch if the shapes of self and src do not match

impl AnyArrayBase<OwnedRepr<()>>

pub fn zeros(dtype: AnyArrayDType, shape: &[usize]) -> Self

Create an array with zeros of dtype and shape shape.

pub fn with_zeros_bytes<T>( dtype: AnyArrayDType, shape: &[usize], with: impl FnOnce(&mut [u8]) -> T, ) -> (Self, T)

Create an array with zeros of dtype and shape shape, and provide mutable access to the bytes of the array.

The array is created to be contiguous and in standard order, i.e. its element order in memory corresponds to the logical order of the array’s elements.

pub fn into_cow(self) -> AnyCowArray<'static>

Returns an owned copy-on-write array.