Trait numcodecs::Codec

pub trait Codec:
    'static
    + Send
    + Sync
    + Clone {
    type Error: 'static + Send + Sync + Error;

    // Required methods
    fn encode(&self, data: AnyCowArray<'_>) -> Result<AnyArray, Self::Error>;
    fn decode(&self, encoded: AnyCowArray<'_>) -> Result<AnyArray, Self::Error>;
    fn decode_into(
        &self,
        encoded: AnyArrayView<'_>,
        decoded: AnyArrayViewMut<'_>,
    ) -> Result<(), Self::Error>;
    fn get_config<S: Serializer>(
        &self,
        serializer: S,
    ) -> Result<S::Ok, S::Error>;
}

Compression codec that encodes and decodes numeric n-dimensional arrays.

Required Associated Types

type Error: 'static + Send + Sync + Error

Error type that may be returned during encodeing and decodeing.

Required Methods

fn encode(&self, data: AnyCowArray<'_>) -> Result<AnyArray, Self::Error>

Encodes the data and returns the result.

Errors

Errors if encoding the buffer fails.

fn decode(&self, encoded: AnyCowArray<'_>) -> Result<AnyArray, Self::Error>

Decodes the encoded data and returns the result.

Errors

Errors if decoding the buffer fails.

fn decode_into( &self, encoded: AnyArrayView<'_>, decoded: AnyArrayViewMut<'_>, ) -> Result<(), Self::Error>

Decodes the encoded data and writes the result into the provided decoded output.

The output must have the correct type and shape.

Errors

Errors if decoding the buffer fails.

fn get_config<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error>

Serializes the configuration parameters for this codec.

The config must include an id field with the DynCodecType::codec_id. The config must be compatible with JSON encoding.

Errors

Errors if serializing the codec configuration fails.

Object Safety

This trait is not object safe.